Refactored add-tags to routes.
