+++ src/index.js
@@ -38,7 +38,7 @@ console.log(filterByEnvironment(...args)); // eslint-disable-line no-console
 sourcemap.install();
 
 // Shared serverModel
-const serverModel = new falcor.Model({
+export const serverModel = new falcor.Model({
   source: new FalcorRouter({ maxPaths: 1000 * 1000 }),
   // maxSize is 400 MB in production and 80 MB when in development or staging mode
   maxSize: filterByEnvironment(400 * 1000 * 1000, 80 * 1000 * 1000),
